backend/pyproject.toml
[project]
name = "emesa-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11,<3.14"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "emesa-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
  python = ">=3.11,<3.14"
  mysqlclient = ">=2.2.7,<3.0.0"
django = ">=5.2.4,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
django-cors-headers = ">=4.7.0,<5.0.0"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
gunicorn = "20.1.0"
pillow = "^11.3.0"

[tool.poetry.dev-dependencies]
